fix zig build check and test
I'm not totally sure what is going on here.  I think running zig build
check ran on the web backend with a native target, but now that we are
more specific about the flags for stb_image, it failed.

I fixed that by disabling an extra stb_image feature when not using
libc, but unsure that is the "right" fix.

Maybe we should compile the web backend checks and tests with a specific
target?